feat(FTDA): Replace create task method from ls to server
