--- shared/redux/rtk-apis/tasksAPI.types.ts
@@ -1,17 +1,0 @@
-import { EPriority, ETaskStatus } from "@/shared/typedefs/enums";
-
-export type TCreateTaskDto = {
-  title: string;
-  description?: string;
-  dueDate?: Date | null;
-  priority: EPriority;
-  status: ETaskStatus;
-};
-
-export type TUpdateTaskDto = {
-  title?: string;
-  description?: string;
-  dueDate?: Date | null;
-  priority?: EPriority;
-  status?: ETaskStatus;
-};
