+++ shared/typedefs/types.ts
@@ -13,3 +13,19 @@ export type TLoginCredentials = {
   email: string;
   password: string;
 };
+
+export type TCreateTaskDto = {
+  title: string;
+  description?: string;
+  dueDate?: Date | null;
+  priority: EPriority;
+  status: ETaskStatus;
+};
+
+export type TUpdateTaskDto = {
+  title?: string;
+  description?: string;
+  dueDate?: Date | null;
+  priority?: EPriority;
+  status?: ETaskStatus;
+};
